core/requirements.txt
Pillow>=6.1.0
